requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API clients
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Job scheduling
APScheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
structlog>=23.1.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
ipython>=8.14.0
